Adding more random bytes to UUID. Supports CN, AN, CE. Draws edges properly now.

+++ nervous/BLESensor.swift
@@ -46,8 +46,8 @@ class BLESensor {
             let date = NSDate()
             let beaconSensor = SensorUpload.builder()
             
-            beaconSensor.huuid = nvm.getHUUID() //phone huuid
-            beaconSensor.luuid = nvm.getLUUID() //phone luuid
+            beaconSensor.huuid = UInt64(nvm.getHUUID()) //phone huuid
+            beaconSensor.luuid = UInt64(nvm.getLUUID()) //phone luuid
 
             beaconSensor.uploadTime = UInt64(currentTime.timeIntervalSince1970*1000)
             beaconSensor.sensorId = 0x000000000000000B
